Cleaned up methods, split up used and unused methods into separate files
